improve nuker performance

+++ Memory/Hooks.cpp
@@ -307,6 +307,7 @@ void Hooks::GameMode_startDestroyBlock(C_GameMode* a, vec3_ti* a2, uint8_t face,
 			vec3_ti tempPos;
 
 			const int range = nukerModule->getNukerRadius();
+			C_BlockSource* region = g_Data.getLocalPlayer()->region;
 
 			for (int x = -range; x < range; x++) {
 				for (int y = -range; y < range; y++) {
@@ -314,8 +315,9 @@ void Hooks::GameMode_startDestroyBlock(C_GameMode* a, vec3_ti* a2, uint8_t face,
 						tempPos.x = a2->x + x;
 						tempPos.y = a2->y + y;
 						tempPos.z = a2->z + z;
-						if (tempPos.y > 0)
+						if (tempPos.y > 0 && (*(region->getBlock(tempPos)->blockLegacy))->blockId != 0) {
 							a->destroyBlock(&tempPos, face);
+						}
 					}
 				}
 			}
